Final GUI Version of Flash Flow
Due to some time constraints and the amount of re-coding, we had to remove a couple of features for this final version to meet the project deadline. Overall, I had a blast developing this project, and I'm excited about my future projects; maybe I'll keep developing this in my spare time. 
Removed: 
The ability to view and remove flashcards from decks
Usernames